Big Update
Can parse the Vector files and store the arrays but the problem is that the output is always 4. This could be the issue with Audrey's code but it also could be my parser not working.

Please make or I will make getter methods for the gates and Ouput Nmaes for future debugging.